#151 refactor settings -> etc
