--- app/src/main/res/drawable/setting_blue.xml
@@ -1,9 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="28dp"
-    android:height="30dp"
-    android:viewportWidth="28"
-    android:viewportHeight="30">
-  <path
-      android:pathData="M23.32,9.55L21.43,8.92L22.32,7.14C22.41,6.954 22.44,6.744 22.406,6.54C22.372,6.336 22.276,6.147 22.13,6L20,3.87C19.852,3.722 19.662,3.624 19.455,3.59C19.249,3.556 19.038,3.587 18.85,3.68L17.07,4.57L16.44,2.68C16.374,2.483 16.247,2.312 16.079,2.19C15.91,2.068 15.708,2.002 15.5,2H12.5C12.29,1.999 12.086,2.065 11.915,2.187C11.745,2.309 11.617,2.481 11.55,2.68L10.92,4.57L9.14,3.68C8.954,3.59 8.744,3.56 8.54,3.594C8.336,3.628 8.147,3.724 8,3.87L5.87,6C5.722,6.148 5.624,6.338 5.59,6.544C5.556,6.751 5.587,6.963 5.68,7.15L6.57,8.93L4.68,9.56C4.483,9.627 4.312,9.753 4.19,9.921C4.068,10.09 4.002,10.292 4,10.5V13.5C3.999,13.71 4.065,13.914 4.187,14.085C4.309,14.255 4.481,14.383 4.68,14.45L6.57,15.08L5.68,16.86C5.59,17.046 5.56,17.256 5.594,17.46C5.628,17.664 5.724,17.853 5.87,18L8,20.13C8.148,20.278 8.338,20.376 8.544,20.41C8.751,20.444 8.963,20.413 9.15,20.32L10.93,19.43L11.56,21.32C11.627,21.519 11.755,21.691 11.925,21.813C12.096,21.935 12.3,22 12.51,22H15.51C15.72,22 15.924,21.935 16.095,21.813C16.265,21.691 16.393,21.519 16.46,21.32L17.09,19.43L18.87,20.32C19.055,20.408 19.263,20.437 19.465,20.403C19.667,20.369 19.854,20.274 20,20.13L22.13,18C22.278,17.852 22.376,17.662 22.41,17.455C22.444,17.249 22.413,17.038 22.32,16.85L21.43,15.07L23.32,14.44C23.517,14.373 23.688,14.247 23.81,14.079C23.932,13.91 23.999,13.708 24,13.5V10.5C24,10.29 23.935,10.086 23.813,9.915C23.691,9.745 23.519,9.617 23.32,9.55ZM22,12.78L20.8,13.18C20.524,13.269 20.271,13.418 20.058,13.615C19.845,13.812 19.678,14.053 19.567,14.322C19.457,14.59 19.406,14.879 19.419,15.169C19.431,15.459 19.507,15.742 19.64,16L20.21,17.14L19.11,18.24L18,17.64C17.744,17.512 17.463,17.441 17.176,17.431C16.89,17.421 16.605,17.473 16.34,17.583C16.076,17.693 15.838,17.859 15.643,18.069C15.448,18.279 15.3,18.528 15.21,18.8L14.81,20H13.22L12.82,18.8C12.731,18.524 12.582,18.271 12.385,18.058C12.188,17.845 11.947,17.678 11.678,17.567C11.41,17.457 11.121,17.406 10.831,17.419C10.541,17.431 10.258,17.507 10,17.64L8.86,18.21L7.76,17.11L8.36,16C8.493,15.742 8.569,15.459 8.581,15.169C8.594,14.879 8.543,14.59 8.433,14.322C8.322,14.053 8.155,13.812 7.942,13.615C7.729,13.418 7.476,13.269 7.2,13.18L6,12.78V11.22L7.2,10.82C7.476,10.731 7.729,10.582 7.942,10.385C8.155,10.188 8.322,9.947 8.433,9.678C8.543,9.41 8.594,9.121 8.581,8.831C8.569,8.541 8.493,8.258 8.36,8L7.79,6.89L8.89,5.79L10,6.36C10.258,6.493 10.541,6.569 10.831,6.581C11.121,6.594 11.41,6.543 11.678,6.433C11.947,6.322 12.188,6.155 12.385,5.942C12.582,5.729 12.731,5.476 12.82,5.2L13.22,4H14.78L15.18,5.2C15.269,5.476 15.418,5.729 15.615,5.942C15.812,6.155 16.053,6.322 16.322,6.433C16.59,6.543 16.879,6.594 17.169,6.581C17.459,6.569 17.742,6.493 18,6.36L19.14,5.79L20.24,6.89L19.64,8C19.512,8.256 19.441,8.537 19.431,8.824C19.421,9.11 19.473,9.395 19.583,9.66C19.693,9.924 19.859,10.162 20.069,10.357C20.279,10.552 20.528,10.7 20.8,10.79L22,11.19V12.78ZM14,8C13.209,8 12.436,8.235 11.778,8.674C11.12,9.114 10.607,9.738 10.304,10.469C10.002,11.2 9.923,12.004 10.077,12.78C10.231,13.556 10.612,14.269 11.172,14.828C11.731,15.388 12.444,15.769 13.22,15.923C13.996,16.077 14.8,15.998 15.531,15.696C16.262,15.393 16.886,14.88 17.326,14.222C17.765,13.564 18,12.791 18,12C18,10.939 17.579,9.922 16.828,9.172C16.078,8.421 15.061,8 14,8ZM14,14C13.604,14 13.218,13.883 12.889,13.663C12.56,13.443 12.304,13.131 12.152,12.765C12.001,12.4 11.961,11.998 12.038,11.61C12.116,11.222 12.306,10.866 12.586,10.586C12.866,10.306 13.222,10.116 13.61,10.038C13.998,9.961 14.4,10.001 14.765,10.152C15.131,10.304 15.443,10.56 15.663,10.889C15.883,11.218 16,11.604 16,12C16,12.53 15.789,13.039 15.414,13.414C15.039,13.789 14.53,14 14,14Z"
-      android:fillColor="#22A7CC"/>
-</vector>
