+++ app/src/main/res/menu/main_menu.xml
@@ -20,7 +20,7 @@
         android:icon="@drawable/menu_selector_mypage_color" />
 
     <item
-        android:id="@id/setting_nav"
+        android:id="@id/etc_nav"
         android:title="@string/etc"
         android:icon="@drawable/menu_selector_etc_color" />
 </menu>
