+++ app/src/main/res/navigation/main_nav.xml
@@ -6,7 +6,7 @@
 
 
     <include app:graph="@navigation/mypage_nav" />
-    <include app:graph="@navigation/setting_nav" />
+    <include app:graph="@navigation/etc_nav" />
     <include app:graph="@navigation/intro_nav" />
     <include app:graph="@navigation/home_nav" />
     <include app:graph="@navigation/news_nav" />
