+++ core/ui/src/main/res/values/attrs.xml
@@ -34,7 +34,7 @@
         </attr>
     </declare-styleable>
 
-    <declare-styleable name="SettingBar">
+    <declare-styleable name="EtcBar">
         <attr name="title" />
         <attr name="showArrow" format="boolean" />
     </declare-styleable>
