+++ feature/etc/build.gradle.kts
@@ -1,12 +1,10 @@
 plugins {
-
     id("mediproject.android.feature")
-    id(libs.plugins.kapt.get().pluginId)
-
 }
 
 android {
-    namespace = "com.android.mediproject.feature.favoritemedicine"
+    namespace = "com.android.mediproject.feature.etc`"
+
     buildFeatures {
         buildConfig = true
     }
@@ -17,15 +15,12 @@ hilt {
 }
 
 dependencies {
-
     implementation(project(":core:ui"))
     implementation(project(":core:common"))
-    implementation(project(":core:domain"))
-    implementation(project(":core:model"))
     implementation(libs.bundles.glides)
     kapt(libs.bundles.glides.kapt)
     implementation(libs.bundles.kotlins)
     implementation(libs.bundles.materials)
     implementation(libs.bundles.uiAndroidx)
-    implementation(libs.android.flexbox)
+
 }
