+++ feature/etc/src/main/AndroidManifest.xml
