+++ feature/etc/src/main/res/values/string.xml
