--- feature/setting/build.gradle.kts
@@ -1,27 +1,0 @@
-plugins {
-
-    id("mediproject.android.feature")
-}
-
-android {
-    namespace = "com.android.mediproject.feature.setting"
-
-    buildFeatures {
-        buildConfig = true
-    }
-}
-
-hilt {
-    enableAggregatingTask = true
-}
-
-dependencies {
-    implementation(project(":core:ui"))
-    implementation(project(":core:common"))
-    implementation(libs.bundles.glides)
-    kapt(libs.bundles.glides.kapt)
-    implementation(libs.bundles.kotlins)
-    implementation(libs.bundles.materials)
-    implementation(libs.bundles.uiAndroidx)
-
-}
