--- feature/setting/src/main/res/navigation/setting_nav.xml
@@ -1,13 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<navigation xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:id="@+id/setting_nav"
-    app:startDestination="@id/settingFragment">
-
-    <fragment
-        android:id="@+id/settingFragment"
-        android:name="com.android.mediproject.feature.setting.SettingFragment"
-        android:label="SettingFragment" />
-
-    <deepLink app:uri="medilens://main/mypage_nav" />
-</navigation>
